[BG-1057] Phing: always run "npm prune" before "npm install" to remove packages that are not required anymore
"npm install" itself does not remove installed but required packages.
Adding "npm prune" solves problem that we had with styledocco and grunt@1.0. Styledocco requires grunt@0.4. We used styledocco back when we had grunt@0.4. Then we stopped to use styledocco so we removed it from packages.json. Then we updated grunt to grunt@1.0. But as styledocco was still present in node_modules, "npm install" failed because styledocco could not be left installed along with grunt@1.0.