GridView: add type hinting annotation because Twig_TemplateInterface does not have getParent() method that Twig_Template has
I guess we should not run into any other implementation of Twig_TemplateInterface than Twig_Template.
Instead of type-hinting we could throw an exception when $template is not an instance of Twig_Template but we get the similar result when we get a runtime exception for the missing method too.