CLI: new command "shopsys:database:create" for more comfortable DB creation
This command takes care of creating the database including required
DB extensions and collations (that are operating system specific, unfortunately).
Therefore it replaces DbCollationsDataFixture and part of DbFunctionsDataFixture.
When a locale is not supported by the operating system the command
explains the situation and links to the documentation.

It also interactively asks for DB superuser credentials in order to perform
all the actions so it is no longer needed to temporarily put superuser credentials
into app/config/parameters.yml.

The command is designed to be run only during first creation of the database
but as it uses "IF NOT EXISTS" commands, it can be manually run on existing
database in order to create new DB extensions or collations, too.

Resolves: https://git.shopsys-framework.com/shopsys/shopsys-framework/issues/7
Resolves: https://git.shopsys-framework.com/shopsys/shopsys-framework/issues/8