[BG-1759] feeds: increase batch size to speed-up generating feeds
Commit d20d3e2153110b975adf2d11238347231fb362e1 adds a workaround for
a problem with inconsistent visibility of products and categories.
The commit adds product visibility recalculation into generateFeedBatch()
method which means that the visibility is not recalculated just once
before start of generating feeds but it is recalculated before every
batch. This is quite a big overhead but unfortunatelly the API of
IteratedCronModuleInterface does not allow to calculate the visibility
just once.
Therefore I decided to increase the batch size from 200 to 1000 which
basically decreases the number of visibility recalculations five times.
This makes the performance tests to pass again.
Note: We need to find a proper solution for the problem that the
visibility of Products and Categories can be sometimes inconsistent.